Fix Panel Register Dependency (#23)
The MaterialSetsPanel requires the LodAndMaterialSetsPanel to be registered first. The file lod_and_material_sets_panel.py nicely delegates this to the LODPanel (in lod_panel.py) However this dependency is not properly enforced. On Windows the os.walk implementation seems to nicely sort the files however this is NOT the case on macOS and you get whatever order of files. Since the current implementation seems to rely on the Windows platform which does the sorting for you, to be platform independent we must sort the list of files from os.walk().

---------

Co-authored-by: Onno Hommes <ohommes@apple.com>